+++ lib/cuprum/collections/collection.rb
@@ -4,27 +4,24 @@ require 'cuprum/command_factory'
 
 require 'cuprum/collections'
 require 'cuprum/collections/relation'
+require 'cuprum/collections/relations/options'
+require 'cuprum/collections/relations/parameters'
+require 'cuprum/collections/relations/primary_keys'
+require 'cuprum/collections/relations/scope'
 require 'cuprum/collections/scopes/all_scope'
 
 module Cuprum::Collections
   # Provides a base implementation for collections.
   class Collection < Cuprum::CommandFactory
-    include Cuprum::Collections::Relation::Parameters
-    include Cuprum::Collections::Relation::PrimaryKeys
+    include Cuprum::Collections::Relations::Options
+    include Cuprum::Collections::Relations::Parameters
+    include Cuprum::Collections::Relations::PrimaryKeys
+    include Cuprum::Collections::Relations::Scope
 
     # Error raised when trying to call an abstract collection method.
     class AbstractCollectionError < StandardError; end
 
-    IGNORED_PARAMETERS = %i[
-      entity_class
-      name
-      query
-      qualified_name
-      singular_name
-    ].freeze
-    private_constant :IGNORED_PARAMETERS
-
-    # @overload initialize(entity_class: nil, name: nil, qualified_name: nil, singular_name: nil, **options)
+    # @!method initialize(entity_class: nil, name: nil, qualified_name: nil, singular_name: nil, **options)
     #   @param entity_class [Class, String] the class of entity represented by
     #     the relation.
     #   @param name [String] the name of the relation.
@@ -34,29 +31,11 @@ module Cuprum::Collections
     #
     #   @option options primary_key_name [String] the name of the primary key
     #     attribute. Defaults to 'id'.
-    #   @option primary_key_type [Class, Stannum::Constraint] the type of
-    #     the primary key attribute. Defaults to Integer.
-    def initialize(**parameters) # rubocop:disable Metrics/MethodLength
-      super()
-
-      relation_params = resolve_parameters(parameters)
-      @entity_class   = relation_params[:entity_class]
-      @name           = relation_params[:name]
-      @plural_name    = relation_params[:plural_name]
-      @qualified_name = relation_params[:qualified_name]
-      @singular_name  = relation_params[:singular_name]
-
-      @scope   =
-        if parameters.key?(:scope)
-          default_scope.and(parameters[:scope])
-        else
-          default_scope
-        end
-      @options = ignore_parameters(**parameters)
-    end
-
-    # @return [Hash<Symbol>] additional options for the collection.
-    attr_reader :options
+    #   @option options primary_key_type [Class, Stannum::Constraint] the type
+    #     of the primary key attribute. Defaults to Integer.
+    #   @option options scope
+    #     [Cuprum::Collections::Scopes::Base, Hash, Proc, nil] the configured
+    #     scope for the relation.
 
     # @param other [Object] The object to compare.
     #
@@ -99,20 +78,7 @@ module Cuprum::Collections
         'subclass and implement the #query method.'
     end
 
-    # @return [Cuprum::Collections::Scopes::Base] the configured scope for the
-    #   collection.
-    def scope
-      @scope ||= default_scope
-    end
-
-    # Returns a copy of the collection that merges the given scope.
-    def with_scope(other_scope)
-      dup.tap { |copy| copy.scope = scope.and(other_scope) }
-    end
-
     protected
-
-    attr_writer :scope
 
     def comparable_options
       @comparable_options ||= {
@@ -125,20 +91,6 @@ module Cuprum::Collections
         singular_name:,
         **options
       }
-    end
-
-    private
-
-    def default_scope
-      Cuprum::Collections::Scopes::AllScope.new
-    end
-
-    def ignore_parameters(**parameters)
-      parameters.except(*ignored_parameters)
-    end
-
-    def ignored_parameters
-      @ignored_parameters ||= Set.new(IGNORED_PARAMETERS)
     end
   end
 end